Streamlit/streamlit-map-template/requirements.txt
--find-links=https://girder.github.io/large_image_wheels GDAL
geopandas
leafmap
owslib
streamlit
